llmService/requirements.txt
fastapi
uvicorn
sse-starlette
python-dotenv
openai  # If you're using OpenAI
google-generativeai
PyJWT
asyncpg